refactor: EmailAddr.Request -> UnverifiedEmailAddr.
