+++ backend/src/seed/factories/index.js
@@ -9,7 +9,7 @@ import createTag from './tags.js'
 import createSocialMedia from './socialMedia.js'
 import createLocation from './locations.js'
 import createEmailAddress from './emailAddresses.js'
-import createEmailAddressRequests from './emailAddressRequests.js'
+import createUnverifiedEmailAddresss from './unverifiedEmailAddresses.js'
 
 export const seedServerHost = 'http://127.0.0.1:4001'
 
@@ -33,7 +33,7 @@ const factories = {
   SocialMedia: createSocialMedia,
   Location: createLocation,
   EmailAddress: createEmailAddress,
-  EmailAddressRequest: createEmailAddressRequests,
+  UnverifiedEmailAddress: createUnverifiedEmailAddresss,
 }
 
 export const cleanDatabase = async (options = {}) => {
